pyproject.toml
[tool.poetry]
name = "konan-sdk"
version = "0.0.0"
description = "Python SDK for Konan's API"
authors = ["Synapse Analytics"]
license = "MIT"
readme = "README.md"
documentation = "https://docs.konan.ai"
repository = "https://github.com/SynapseAnalytics/konan-sdk"

[tool.poetry.urls]
"Konan Website" = "https://konan.ai"

[tool.poetry.dependencies]
python = "^3.7"
requests = "^2.26.0"
PyJWT = { extras = ["crypto"], version = "^2.1.0" }
loguru = "^0.5.3"
pydantic = "^1.8.2"
fastapi = "^0.70"
fastapi-utils = "^0.2.1"

[tool.poetry.dev-dependencies]
flake8 = "^3.9.2"
pep8-naming = "^0.12.1"
pytest = "^6.2.4"
python-semantic-release = "^7.19.2"
PyYAML = "^6.0"
pytz = "^2021.3"
Sphinx = "^4.4.0"
sphinx-rtd-theme = "^1.0.0"

[tool.semantic_release]
branch = "main"
version_variable = [
    "konan_sdk/__init__.py:__version__",
    "docs/source/conf.py:version",
    "pyproject.toml:version",
]
version_toml = [
    "pyproject.toml:tool.poetry.version"
]
version_pattern = [
    # "docs/requirements.txt:konan_sdk=={version}",
]
commit_author = "github-actions <actions@github.com>"
upload_to_repository = true
upload_to_release = true
build_command = "pip install poetry && poetry build"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
